+++ src/error.rs
@@ -25,7 +25,7 @@ impl Error {
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "error performing request: {}", self.message)
+        write!(f, "error: {}", self.message)
     }
 }